Refactor link modal to improve the performance of documents with many pages
The link modal creates a dropdown with an option for every page number
so for large documents (thousands of pages) this can take quite a while
to render.

This change updates the behavior so that you can type the page number
instead and the link button is disabled if the page is out of bounds.