[Bugfix] Iterate Through Fields for Digital Signatures (#3503)
* Iterate through Fields for DigSigs

The previous implementation of Digital Signature Verification UI for
WebViewer iterated through Widgets to determine which Fields in the
document to verify for Digital Signatures. However, per @rastko, there
may be cases where a Digital Signature occurs on a Field and has no
associated Widget.

As such, refactor the Digital Signature Validation code to iterate
through Fields in the documents, rather than the Widgets in the document
to ensure Documents with no Widgets, but only Fields, can still have
their Digital Signatures verified via the UI logic

* Add todo for DigitalSignatureFieldIterator check

* Remove errant console.log calls

* Retrieve field names from PDFNet

Because we rely on PDFNet for Digital Signature Verification logic, for
the sake of consistency, we should also rely on PDFNet to inform
WebViewer which signature fields contain a cryptographic signature, and
which fields are empty

Therefore, refactor `setVerificationResult` and `getVerificationResult`
to not take an argument indicating which field names to find in the
document, but instead, use PDFNet's FieldIterator method to iterate
through all fields, filter for just signature fields, and return this
information to WebViewer, and have WebViewer render the relevant UI
components based on what PDFNet returns

* Ensure spinner is active when validating

* Change unsigned field from obj num to field name

Makes it more obvious to the end-user what field has not yet been signed

* Add logic to highlight and jump to unsigned field

* Translate "unsigned signature field"

* Add link class to unsigned field div

Ensures users are aware it is an interactive element

* Add tabIndex and onKeyPress listener

Ensures the element is interactive with the keyboard as well

* Remove else block and fix sig panel lifecycle

In `setVerificationResult`, the else block was redundant because the
`SignaturePanel` component should be dictating when everything is ready
(i.e. certificates & document) to properly allow the logic in
`setVerificationResult` to work correctly

Fix up the `useEffect` and `state` lifecycle in the `SignaturePanel`
component to make it more understandable about the workflow (i.e. make
the `useEffect` that calls `setVerificationResult` depend on both a
`certificate` and `document` state, and invoke `setDocument` when
`documentLoaded` is triggered if the document has not yet been fully
loaded

* Fix linting error space-before-function-paren